First commit
----- warning ------
this is only a test demo, i wanted this up on my git so i can pull it
work on it anywhere.
There are a lot of errors, but and a lot of things that are completely
wrong with the code,
major code refactoring needs to be done to this repo.

if you can help, please do hopefully we can bring a taste of .net
lambdas to java8 :p